backend/requirements.txt
flask
flask_restful
flask_security_too
flask_migrate
flask_sqlalchemy